requirements.txt
Flask==2.2.2
python-dotenv==0.21.1
requests==2.28.2

